+++ python_modules/libraries/dagster-dg/dagster_dg/cli/check.py
@@ -1,7 +1,6 @@
-import contextlib
 import subprocess
 import sys
-from collections.abc import Iterator, Mapping, Sequence
+from collections.abc import Mapping, Sequence
 from pathlib import Path
 from typing import Any, NamedTuple, Optional
 
@@ -10,13 +9,14 @@ from jsonschema import Draft202012Validator, ValidationError
 from yaml.scanner import ScannerError
 
 from dagster_dg.cli.check_utils import error_dict_to_formatted_error
-from dagster_dg.cli.dev import create_temp_workspace_file, format_forwarded_option
+from dagster_dg.cli.dev import format_forwarded_option
 from dagster_dg.cli.shared_options import dg_global_options
+from dagster_dg.cli.utils import create_dagster_cli_cmd
 from dagster_dg.component import RemoteComponentRegistry
 from dagster_dg.component_key import ComponentKey
 from dagster_dg.config import normalize_cli_config
 from dagster_dg.context import DgContext
-from dagster_dg.utils import DgClickCommand, DgClickGroup, exit_with_error, pushd
+from dagster_dg.utils import DgClickCommand, DgClickGroup, pushd
 from dagster_dg.yaml_utils import parse_yaml_with_source_positions
 from dagster_dg.yaml_utils.source_position import (
     LineCol,
@@ -224,9 +224,19 @@ def check_definitions_command(
         *(["--verbose"] if verbose else []),
     ]
 
+    run_cmds = (
+        ["uv", "run", "dagster", "definitions", "validate"]
+        if dg_context.is_project
+        else ["uv", "tool", "run", "dagster", "definitions", "validate"]
+    )
+
     with (
         pushd(dg_context.root_path),
-        create_validate_cmd(dg_context, forward_options) as (cmd_location, cmd, workspace_file),
+        create_dagster_cli_cmd(dg_context, forward_options, run_cmds) as (
+            cmd_location,
+            cmd,
+            workspace_file,
+        ),
     ):
         print(f"Using {cmd_location}")  # noqa: T201
         if workspace_file:  # only non-None deployment context
@@ -239,38 +249,3 @@ def check_definitions_command(
             sys.exit(result.returncode)
 
     click.echo("All definitions loaded successfully.")
-
-
-class CommandArgs(NamedTuple):
-    cmd_location: str
-    cmd: list[str]
-    workspace_file: Optional[str]
-
-
-@contextlib.contextmanager
-def create_validate_cmd(dg_context: DgContext, forward_options: list[str]) -> Iterator[CommandArgs]:
-    val_args = ["definitions", "validate", *forward_options]
-    if dg_context.is_project:
-        # In a code location context, we can just run `dagster definitions validate` directly, using `dagster` from the
-        # code location's environment.
-        cmd = ["uv", "run", "dagster", *val_args]
-        cmd_location = dg_context.get_executable("dagster")
-        with create_temp_workspace_file(dg_context) as temp_workspace_file:
-            yield CommandArgs(
-                cmd_location=str(cmd_location), cmd=cmd, workspace_file=temp_workspace_file
-            )
-        # Still need to get test_validate_command_project_context_success to work
-        # yield CommandArgs(cmd_location=str(cmd_location), cmd=cmd, workspace_file=None)
-    elif dg_context.is_workspace:
-        # In a workspace context, dg validate will construct a temporary
-        # workspace file that points at all defined code locations and invoke:
-        #
-        #     uv tool run --with dagster-webserver dagster definitions validate
-        with create_temp_workspace_file(dg_context) as temp_workspace_file:
-            yield CommandArgs(
-                cmd=["uv", "tool", "run", "dagster", *val_args],
-                cmd_location="ephemeral dagster definitions validate",
-                workspace_file=temp_workspace_file,
-            )
-    else:
-        exit_with_error("This command must be run inside a code location or deployment directory.")
