+++ python_modules/libraries/dagster-dg/dagster_dg/cli/dev.py
@@ -3,21 +3,19 @@ import signal
 import subprocess
 import sys
 import time
-from collections.abc import Iterator, Mapping, Sequence
-from contextlib import contextmanager
+from collections.abc import Mapping, Sequence
 from pathlib import Path
-from tempfile import NamedTemporaryFile
 from typing import Optional, TypeVar
 
 import click
 import psutil
-import yaml
 
 from dagster_dg.cli.shared_options import dg_global_options
+from dagster_dg.cli.utils import create_dagster_cli_cmd
 from dagster_dg.config import normalize_cli_config
 from dagster_dg.context import DgContext
 from dagster_dg.error import DgError
-from dagster_dg.utils import DgClickCommand, exit_with_error, pushd
+from dagster_dg.utils import DgClickCommand, pushd
 
 T = TypeVar("T")
 
@@ -106,40 +104,21 @@ def dev_command(
     ]
 
     # In a project context, we can just run `dagster dev` directly, using `dagster` from the
-    # project's environment.
-    temp_workspace_file_cm = create_temp_workspace_file(dg_context)
-    if dg_context.is_project:
-        cmd_location = dg_context.get_executable("dagster")
-        if dg_context.use_dg_managed_environment:
-            cmd = ["uv", "run", "dagster", "dev", *forward_options]
-        else:
-            cmd = [cmd_location, "dev", *forward_options]
-
+    # code location's environment.
     # In a workspace context, dg dev will construct a temporary
-    # workspace file that points at all defined projects and invoke:
-    #
+    # workspace file that points at all defined code locations and invoke:
     #     uv tool run --with dagster-webserver dagster dev
-    #
-    # The `--with dagster-webserver` is necessary here to ensure that dagster-webserver is
-    # installed in the isolated environment that `uv` will install `dagster` in.
-    # `dagster-webserver` is not a dependency of `dagster` but is required to run the `dev`
-    # command.
-    elif dg_context.is_workspace:
-        cmd = [
-            "uv",
-            "tool",
-            "run",
-            "--with",
-            "dagster-webserver",
-            "dagster",
-            "dev",
-            *forward_options,
-        ]
-        cmd_location = "ephemeral dagster dev"
-    else:
-        exit_with_error("This command must be run inside a project or workspace directory.")
+    run_cmds = (
+        ["uv", "run", "dagster", "dev"]
+        if dg_context.is_project
+        else ["uv", "tool", "run", "--with", "dagster-webserver", "dagster", "dev"]
+    )
 
-    with pushd(dg_context.root_path), temp_workspace_file_cm as workspace_file:
+    with (
+        pushd(dg_context.root_path),
+        create_dagster_cli_cmd(dg_context, forward_options, run_cmds=run_cmds) as cmd_object,
+    ):
+        cmd_location, cmd, workspace_file = cmd_object
         print(f"Using {cmd_location}")  # noqa: T201
         if workspace_file:  # only non-None deployment context
             cmd.extend(["--workspace", workspace_file])
@@ -169,33 +148,6 @@ def dev_command(
             except subprocess.TimeoutExpired:
                 click.secho("`dagster dev` did not terminate in time. Killing it.")
                 uv_run_dagster_dev_process.kill()
-
-
-@contextmanager
-def create_temp_workspace_file(dg_context: DgContext) -> Iterator[str]:
-    with NamedTemporaryFile(mode="w+", delete=True) as temp_workspace_file:
-        entries = []
-        if dg_context.is_project:
-            entries.append(_workspace_entry_for_project(dg_context))
-        elif dg_context.is_workspace:
-            for spec in dg_context.project_specs:
-                project_root = dg_context.root_path / spec.path
-                project_context: DgContext = dg_context.with_root_path(project_root)
-                entries.append(_workspace_entry_for_project(project_context))
-        yaml.dump({"load_from": entries}, temp_workspace_file)
-        temp_workspace_file.flush()
-        yield temp_workspace_file.name
-
-
-def _workspace_entry_for_project(dg_context: DgContext) -> dict[str, dict[str, str]]:
-    entry = {
-        "working_directory": str(dg_context.root_path),
-        "module_name": str(dg_context.code_location_target_module_name),
-        "location_name": dg_context.code_location_name,
-    }
-    if dg_context.use_dg_managed_environment:
-        entry["executable_path"] = str(dg_context.project_python_executable)
-    return {"python_module": entry}
 
 
 def format_forwarded_option(option: str, value: object) -> list[str]:
